+++ app/src/main/java/com/yuganshtyagi/whatsapp_copy/CallsFragment.java
@@ -1,4 +1,4 @@
-package com.yuganshtyagi.whatsapp_copy.Fragment;
+package com.yuganshtyagi.whatsapp_copy;
 
 import android.graphics.Color;
 import android.os.Bundle;
@@ -11,31 +11,28 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.yuganshtyagi.whatsapp_copy.Adapter.ChatAdapter;
-import com.yuganshtyagi.whatsapp_copy.R;
-import com.yuganshtyagi.whatsapp_copy.Utils.SeparatorDecoration;
-
 /**
  * Created by Yugansh Tyagi on 3/21/2018.
  */
 
-public class ChatsFragment extends Fragment {
+public class CallsFragment extends Fragment {
 
     RecyclerView recyclerView;
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.chats_fragment_layout, container, false);
+        View view = inflater.inflate(R.layout.calls_fragment_layout, null, false);
 
         SeparatorDecoration decoration = new SeparatorDecoration(
                 getContext(),
                 Color.parseColor("#EAEAEA"),
                 0.5f);
-        recyclerView = view.findViewById(R.id.chat_rv);
+
+        recyclerView = view.findViewById(R.id.calls_rv);
         recyclerView.setHasFixedSize(true);
         recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
-        recyclerView.setAdapter(new ChatAdapter(getContext()));
+        recyclerView.setAdapter(new CallsAdapter(getContext()));
         recyclerView.addItemDecoration(decoration);
 
         return view;
