+++ app/src/main/java/com/yuganshtyagi/whatsapp_copy/StatusAdapter.java
@@ -1,4 +1,4 @@
-package com.yuganshtyagi.whatsapp_copy.Adapter;
+package com.yuganshtyagi.whatsapp_copy;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
