+++ app/src/main/java/com/yuganshtyagi/whatsapp_copy/StatusFragment.java
@@ -1,4 +1,4 @@
-package com.yuganshtyagi.whatsapp_copy.Fragment;
+package com.yuganshtyagi.whatsapp_copy;
 
 import android.graphics.Color;
 import android.os.Bundle;
@@ -10,10 +10,6 @@ import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
-import com.yuganshtyagi.whatsapp_copy.Adapter.StatusAdapter;
-import com.yuganshtyagi.whatsapp_copy.R;
-import com.yuganshtyagi.whatsapp_copy.Utils.SeparatorDecoration;
 
 /**
  * Created by Yugansh Tyagi on 3/21/2018.
